refactor(print-format-builder): extract RepeaterFieldInspector
